requirements.txt
llama-index
pypdf
python-dotenv
pandas
langchain-community